% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-package.R
\docType{package}
\name{practice-package}
\alias{practice}
\alias{practice-package}
\title{practice: Practice Package}
\description{
This package includes a function that reads all excel files in a folder and merges them in one data frame.
}
\author{
\strong{Maintainer}: Anna Mantsoki \email{anna.mantsoki@finddx.org} (\href{https://orcid.org/xx}{ORCID})

Other contributors:
\itemize{
  \item FIND [copyright holder]
}

}
\keyword{internal}
